refactor(RoomView): share a RoomPlaceholder shell across blocking screens
