+++ app/views/RoomView/components/EncryptedRoom.tsx
@@ -1,18 +1,15 @@
 import { type ReactElement } from 'react';
-import { Linking, StyleSheet, Text, View } from 'react-native';
+import { Linking } from 'react-native';
 import { type NativeStackNavigationProp } from '@react-navigation/native-stack';
 
 import { type ChatsStackParamList } from '../../../stacks/types';
 import { useTheme } from '../../../theme';
-import { CustomIcon } from '../../../containers/CustomIcon';
 import Button from '../../../containers/Button';
-import sharedStyles from '../../Styles';
 import { useMasterDetail } from '../../../lib/hooks/useMasterDetail';
 import { LEARN_MORE_E2EE_URL } from '../../../lib/encryption/constants';
 import I18n from '../../../i18n';
 import { type TNavigation } from '../../../stacks/stackType';
-
-const GAP = 32;
+import { RoomPlaceholder } from './RoomPlaceholder';
 
 export const EncryptedRoom = ({
 	roomName,
@@ -22,7 +19,6 @@ export const EncryptedRoom = ({
 	navigation: NativeStackNavigationProp<ChatsStackParamList & TNavigation, 'RoomView'>;
 }): ReactElement => {
 	const { colors } = useTheme();
-	const styles = useStyle();
 	const isMasterDetail = useMasterDetail();
 
 	const navigate = () => {
@@ -34,65 +30,18 @@ export const EncryptedRoom = ({
 	};
 
 	return (
-		<View style={styles.root} testID='room-view-encrypted-room'>
-			<View style={styles.container}>
-				<View style={styles.textView}>
-					<View style={styles.icon}>
-						<CustomIcon name='encrypted' size={42} color={colors.fontSecondaryInfo} />
-					</View>
-					<Text style={styles.title}>{I18n.t('encrypted_room_title', { room_name: roomName.slice(0, 30) })}</Text>
-					<Text style={styles.description}>{I18n.t('encrypted_room_description')}</Text>
-				</View>
-				<Button title={I18n.t('Enter_E2EE_Password')} onPress={navigate} />
-				<Button
-					title={I18n.t('Learn_more')}
-					type='secondary'
-					backgroundColor={colors.surfaceTint}
-					onPress={() => Linking.openURL(LEARN_MORE_E2EE_URL)}
-				/>
-			</View>
-		</View>
+		<RoomPlaceholder
+			icon='encrypted'
+			title={I18n.t('encrypted_room_title', { room_name: roomName.slice(0, 30) })}
+			description={I18n.t('encrypted_room_description')}
+			testID='room-view-encrypted-room'>
+			<Button title={I18n.t('Enter_E2EE_Password')} onPress={navigate} />
+			<Button
+				title={I18n.t('Learn_more')}
+				type='secondary'
+				backgroundColor={colors.surfaceTint}
+				onPress={() => Linking.openURL(LEARN_MORE_E2EE_URL)}
+			/>
+		</RoomPlaceholder>
 	);
-};
-
-const useStyle = () => {
-	const { colors } = useTheme();
-	const styles = StyleSheet.create({
-		root: {
-			flex: 1,
-			backgroundColor: colors.surfaceRoom
-		},
-		container: {
-			flex: 1,
-			marginHorizontal: 24,
-			justifyContent: 'center'
-		},
-		textView: { alignItems: 'center' },
-		icon: {
-			width: 58,
-			height: 58,
-			borderRadius: 30,
-			marginBottom: GAP,
-			backgroundColor: colors.surfaceNeutral,
-			alignItems: 'center',
-			justifyContent: 'center'
-		},
-		title: {
-			...sharedStyles.textBold,
-			fontSize: 24,
-			lineHeight: 32,
-			textAlign: 'center',
-			color: colors.fontTitlesLabels,
-			marginBottom: GAP
-		},
-		description: {
-			...sharedStyles.textRegular,
-			fontSize: 16,
-			lineHeight: 24,
-			textAlign: 'center',
-			color: colors.fontDefault,
-			marginBottom: GAP
-		}
-	});
-	return styles;
 };
